+++ tokenized_midi_datasets/TokenizedMidiDataset.py
@@ -1,22 +1,19 @@
 import json
-from typing import List
 
 import datasets
 import numpy as np
 import fortepyan as ff
-from tqdm import tqdm
-from midi_tokenizers.one_time_tokenizer import NoLossTokenizer
-from datasets import Split, Dataset, DatasetInfo, BuilderConfig, GeneratorBasedBuilder
+from datasets import Split, DatasetInfo, BuilderConfig, GeneratorBasedBuilder
 
 _DESC = """
-Dataset with midi files, tokenzied using NoLossTokenizer, with records of equal size.
+Dataset with midi files, tokenzied using MidiTokenizer with records of equal size.
 """
 
 
-class ExponentialTimeTokenConfig(BuilderConfig):
+class TokenizedMidiDatasetConfig(BuilderConfig):
     def __init__(
         self,
-        base_dataset_name: str = "roszcz/maestro-sustain-v2",
+        base_dataset_name: str = "roszcz/maestro-v1-sustain",
         extra_datasets: list[str] = [],
         sequence_length: int = 64,
         sequence_step: int = 42,
@@ -34,61 +31,61 @@ class ExponentialTimeTokenConfig(BuilderConfig):
         self.tokenizer_parameters = tokenizer_parameters
 
 
-class ExponentialTimeTokenDataset(GeneratorBasedBuilder):
+class TokenizedMidiDataset(GeneratorBasedBuilder):
     def _info(self) -> DatasetInfo:
         return DatasetInfo(description=_DESC)
 
-    BUILDER_CONFIG_CLASS = ExponentialTimeTokenConfig
+    BUILDER_CONFIG_CLASS = TokenizedMidiDatasetConfig
     BUILDER_CONFIGS = [
-        ExponentialTimeTokenConfig(
+        TokenizedMidiDatasetConfig(
             base_dataset_name="roszcz/maestro-sustain-v2",
             extra_datasets=["roszcz/giant-midi-sustain-v2"],
-            sequence_length=64,
-            sequence_step=16,
+            sequence_length=256,
+            sequence_step=64,
             tokenizer_parameters={"min_time_unit": 0.001, "n_velocity_bins": 32},
             name="giant-short",
         ),
-        ExponentialTimeTokenConfig(
+        TokenizedMidiDatasetConfig(
             base_dataset_name="roszcz/maestro-sustain-v2",
             extra_datasets=[],
-            sequence_length=64,
-            sequence_step=16,
+            sequence_length=256,
+            sequence_step=64,
             tokenizer_parameters={"min_time_unit": 0.001, "n_velocity_bins": 32},
             name="basic-short",
         ),
-        ExponentialTimeTokenConfig(
+        TokenizedMidiDatasetConfig(
             base_dataset_name="roszcz/maestro-sustain-v2",
             extra_datasets=["roszcz/giant-midi-sustain-v2"],
-            sequence_length=256,
-            sequence_step=32,
+            sequence_length=512,
+            sequence_step=64,
             tokenizer_parameters={"min_time_unit": 0.001, "n_velocity_bins": 32},
             name="giant-mid",
         ),
-        ExponentialTimeTokenConfig(
+        TokenizedMidiDatasetConfig(
             base_dataset_name="roszcz/maestro-sustain-v2",
             extra_datasets=[],
-            sequence_length=256,
-            sequence_step=32,
+            sequence_length=512,
+            sequence_step=64,
             tokenizer_parameters={"min_time_unit": 0.001, "n_velocity_bins": 32},
             name="basic-mid",
         ),
-        ExponentialTimeTokenConfig(
+        TokenizedMidiDatasetConfig(
             base_dataset_name="roszcz/maestro-sustain-v2",
             extra_datasets=["roszcz/giant-midi-sustain-v2"],
-            sequence_length=512,
-            sequence_step=32,
+            sequence_length=1024,
+            sequence_step=64,
             tokenizer_parameters={"min_time_unit": 0.001, "n_velocity_bins": 32},
             name="giant-long",
         ),
-        ExponentialTimeTokenConfig(
+        TokenizedMidiDatasetConfig(
             base_dataset_name="roszcz/maestro-sustain-v2",
             extra_datasets=[],
-            sequence_length=512,
-            sequence_step=32,
+            sequence_length=1024,
+            sequence_step=64,
             tokenizer_parameters={"min_time_unit": 0.001, "n_velocity_bins": 32},
             name="basic-long",
         ),
-        ExponentialTimeTokenConfig(
+        TokenizedMidiDatasetConfig(
             base_dataset_name="roszcz/maestro-sustain-v2",
             extra_datasets=[],
             sequence_length=512,
@@ -99,7 +96,7 @@ class ExponentialTimeTokenDataset(GeneratorBasedBuilder):
     ]
     DEFAULT_CONFIG_NAME = "basic-mid"
 
-    def _split_generators(self, dl_manager: datasets.DownloadManager) -> List[datasets.SplitGenerator]:
+    def _split_generators(self, dl_manager: datasets.DownloadManager) -> list[datasets.SplitGenerator]:
         base = datasets.load_dataset(self.config.base_dataset_name)
 
         other_datasets = [datasets.load_dataset(path, split="train") for path in self.config.extra_datasets]
@@ -176,16 +173,3 @@ class ExponentialTimeTokenDataset(GeneratorBasedBuilder):
             start = finish
 
         return pieces
-
-    def _generate_examples(self, dataset_shards: list[Dataset]):
-        self.tokenizer = NoLossTokenizer(**self.config.tokenizer_parameters)
-        for dataset in dataset_shards:
-            for it, record in tqdm(enumerate(dataset), total=len(dataset)):
-                piece = ff.MidiPiece.from_huggingface(dict(record))
-
-                pieces = self.filter_pauses(piece)
-                chopped_sequences = sum([self.piece_to_records(piece) for piece in pieces], [])
-
-                for jt, sequence in enumerate(chopped_sequences):
-                    key = f"{it}_{jt}"
-                    yield key, sequence
